another pass in qnx. fixed some modules and disabled others.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@1548 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/components/laserfeatureextractor/section.h
@@ -22,7 +22,7 @@ public:
     SectionEl() {}
     SectionEl( double r, double b ) { r_ = r; b_ = b; x_ = r*std::cos(b); y_ = r*std::sin(b); };
 
-    void setXY( double x, double y ) { r_ = hypotf(y,x); b_ = std::atan2(y,x); x_ = x; y_ = y; };
+    void setXY( double x, double y ) { r_ = std::hypot(y,x); b_ = std::atan2(y,x); x_ = x; y_ = y; };
   
     double x() const { return x_; };
     double y() const { return y_; };
